Add changes required for arm64 vs tools to work
Various compilation and typo fixes to support compiling the code
in Visual Studio.

Related work items: #17155713

+++ ArmPkg/Library/ArmArchTimerLib/ArmArchTimerLib.c
@@ -201,7 +201,7 @@ GetTimeInNanoSecond (
   UINT64  Remainder;
   UINT64  TimerFreq;
 
-  TimerFreq = GetPlatformTimerFreq ();
+  TimerFreq = (UINT32)GetPlatformTimerFreq ();  // MU_CHANGE - ARM64 VS change
   //
   //          Ticks
   // Time = --------- x 1,000,000,000
