+++ ArmPlatformPkg/PeilessSec/PeilessSec.c
@@ -116,7 +116,7 @@ SecMain (
   BuildStackHob (StackBase, StacksSize);
 
   // TODO: Call CpuPei as a library
-  BuildCpuHob (ArmGetPhysicalAddressBits (), PcdGet8 (PcdPrePiCpuIoSize));
+  BuildCpuHob ((UINT8)ArmGetPhysicalAddressBits (), PcdGet8 (PcdPrePiCpuIoSize));
 
   if (ArmIsMpCore ()) {
     // Only MP Core platform need to produce gArmMpCoreInfoPpiGuid
